+++ demo/pages/common/navigation_bar.py
@@ -2,9 +2,9 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from pages import PyPIComponent, PyPIPage
+from pages import PyPIComponent
 
-from pomcorn import locators
+from pomcorn import Element, locators
 
 if TYPE_CHECKING:
     from pages.help_page import HelpPage
@@ -17,28 +17,16 @@ if TYPE_CHECKING:
 class Navbar(PyPIComponent):
     """Component representing navigation bar in the top of web application."""
 
-    def __init__(
-        self,
-        page: PyPIPage,
-        base_locator: locators.XPathLocator = locators.ClassLocator(
-            class_name="horizontal-menu",
-            # We specify `container` here because the page has several `nav`
-            # tags and several elements with a similar class name.
-            container="nav",
-        ),
-        # If you don't need to wait for the component to become visible during
-        # initialization, set `wait_until_visible` to `False`.
-        wait_until_visible: bool = True,
-    ):
-        super().__init__(page, base_locator, wait_until_visible)
-        self.help_button = self.init_element(
-            # Specifying the argument name is required here, because depending
-            # on the name `init_element` method can modify the passed locator.
-            relative_locator=locators.ElementWithTextLocator(
-                text="Help",
-                element="a",
-            ),
-        )
+    base_locator = locators.ClassLocator(
+        class_name="horizontal-menu",
+        # We specify `container` here because the page has several `nav`
+        # tags and several elements with a similar class name.
+        container="nav",
+    )
+
+    help_button = Element(
+        locator=locators.ElementWithTextLocator(text="Help", element="a"),
+    )
 
     def open_help(self) -> HelpPage:
         """Click on `Help` button and redirect to HelpPage."""