+++ pomcorn/__init__.py
@@ -1,5 +1,6 @@
 from pomcorn.component import Component, ListComponent
-from pomcorn.element import Element, XPathElement
+from pomcorn.descriptors import Element
+from pomcorn.element import XPathElement
 from pomcorn.page import Page
 from pomcorn.web_view import WebView