+++ pomcorn/descriptors/__init__.py
@@ -1,0 +1,3 @@
+from pomcorn.descriptors.element import Element
+
+__all__ = ("Element",)